==> studies/study2/1_study2_data.qmd <==
---
title: "Simulating study 2 data"
author: "Jenna Landy"
---

```{r}
library(MCMCpack)
library(glue)
library(bayesNMF)

data_dir <- "../../data/study2"
```

## Set Parameters
```{r}
reference <- get_cosmic()

G <- 64
Ns <- 1:20
K <- nrow(reference)

Nreps <- 10

cos_threshold <- 0.8
```

```{r}
# set hyperparameters for distribution of m_g ~ NB(r * N, p)
p <- 0.1
mean_mutations_per_signature <- 1000 # r(1-p)/p
r <- mean_mutations_per_signature * p / (1 - p)

hist(rnbinom(10000, size = r*4, prob = p))
```

## Simulate
Each data object is a list containing
- M: data matrix (mutational counts), K mutation types x G samples
- P: generating factors (mutational signatures), K mutation types x N factors
- E: generating weights (exposures), N factors x G samples

```{r}
set.seed(321)
for (N in Ns) {
  for (rep in 1:Nreps) {
    # sample a subset of N reference signatures
    # require no sampled signatures have cosine similarity > cos_threshold
    sigs <- c()
    available_sigs <- 1:ncol(reference)
    while(length(sigs) < N & length(available_sigs) > 0) {
      if (length(sigs) > 0) {
        sim <- pairwise_sim(reference[,sigs], reference[,available_sigs])
        remove <- which(colSums(sim > cos_threshold) > 0)
        available_sigs <- available_sigs[!available_sigs %in% available_sigs[remove]]
      }
      sampled <- sample(available_sigs, size = 1)
      sigs <- c(sigs, sampled)
    }
    if(length(sigs) < N) {
      print(paste0("Warning: only ", length(sigs), " signatures sampled for N = ", N, " rep = ", rep))
      stop("Not enough signatures sampled")
      next
    }
    sim <- pairwise_sim(reference[,sigs], reference[,sigs])
    stopifnot(all(colSums(sim > cos_threshold) == 1)) # the allowed 1 is with themselves

    P <- as.matrix(reference[,sigs])
    if (N == 1) {
      P = matrix(P, ncol = 1)
    }

    # sample m_g = E[# mutations for subject g]
    # scales with number of signatures, so all signatures have 
    # approximately equal power
    m <- rnbinom(G, size = r * N, prob = p)

    # sample contributions Eg and mutational counts Mg
    E <- matrix(nrow = N, ncol = G)
    M <- matrix(nrow = K, ncol = G)
    for (g in 1:G) {
      probs <- rdirichlet(1, rep(1, N))
      E[,g] <- rmultinom(1, size = m[g], prob = probs)
      M[,g] <- rpois(K, P%*%E[,g])
    }

    # save full dataset
    data <- list(
      M = M,
      P = P,
      E = E
    )
    saveRDS(data, file = file.path(
      data_dir, 
      glue("N{N}_G{G}_rep{rep}.rds")
    ))
  }
}
```

Separately, save a csv file of the $M$ data matrix for each simulation (this is needed to run SignatureAnalyzer).
```{r}
files = list.files(data_dir)
files = files[grepl('.rds', files)]
for (file in files) {
  data = readRDS(file.path(data_dir, file))
  rownames(data$M) = rownames(reference)
  write.csv(
    data$M, 
    file.path(data_dir, str_replace(file, '.rds', '.csv'))
  )
}
```
